Update MathModule to use Double instead Float. Fixed #171

+++ src/main/scala/carldata/sf/core/HydrologyModule.scala
@@ -28,22 +28,22 @@ class HydrologyModule extends Runtime {
     * 'u': units: mm or in
     */
 
-  def $manning_flow(n: Float, d: Float, s: Float, u: String): Float = {
+  def $manning_flow(n: Double, d: Double, s: Double, u: String): Double = {
     $manning_velocity(n, d, s, u) * cross_sectional_area(d)
   }
 
-  def $manning_velocity(n: Float, d: Float, s: Float, u: String): Float = {
+  def $manning_velocity(n: Double, d: Double, s: Double, u: String): Double = {
     u match {
-      case "mm" => ((1 / n) * Math.pow(0.25 * d / 1000 , 0.6667) * Math.pow(s, 0.5)).toFloat
-      case _ => ((1.49f / n) * Math.pow(0.25 * d / 12 , 0.6667) * Math.pow(s, 0.5)).toFloat
+      case "mm" => (1 / n) * Math.pow(0.25 * d / 1000, 0.6667) * Math.pow(s, 0.5)
+      case _ => (1.49 / n) * Math.pow(0.25 * d / 12, 0.6667) * Math.pow(s, 0.5)
     }
   }
 
   /**
     * For pipe flowing full
     **/
-  private def cross_sectional_area(d: Float): Float = {
-    (Math.PI * d * d).toFloat / 4
+  private def cross_sectional_area(d: Double): Double = {
+    (Math.PI * d * d) / 4
   }
